backend2/requirements.txt
python-docx==0.8.11
flask>=2.0.0
flask-cors>=3.0.0
werkzeug>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
